\name{aqfig-package}
\alias{aqfig-package}
\alias{aqfig}

\docType{package}

\title{Functions to help draw figures displaying air quality data
  and model output}

\description{This package contains several functions to help users draw
  figures to display air pollution measurements and air quality model
  output.  These include functions to place a color legend to the right
  of a plot, to draw a scatterplot with the points' colors or sizes
  reflecting the value of a response variable, etc.}
