\name{ozone2}
\docType{data}
\alias{ozone2}

\title{Daily maximum ozone concentrations}

\description{This data set gives the daily maximum ozone
  concentrations (in ppb) observed at a subset of 23 Clean Air
  Status and Trends Network (CASTNET) monitoring sites on
  August 10, 2004.  }


\usage{ozone2}

\format{A file in CSV format containing 23 records.}

\source{Hourly ozone data at CASTNET sites can be obtained from EPA's
  Clean Air Status and Trends Network program at 
  http://www.epa.gov/castnet.  Thanks to Steven Howard for his help
  collating and summarizing these hourly data.}


\keyword{datasets}
