\name{aqfig-package}
\alias{aqfig-package}
\alias{aqfig}

\docType{package}

\title{Functions to help draw figures displaying air quality data
  and model output}

\description{This package contains several functions to help users draw
  figures to display air pollution measurements and air quality model
  output.  These include functions to generate figures with appropriate
  dimensions and aspect ratio, to place a legend to the right of a plot,
  to draw a scatterplot with the points' colors or sizes reflecting the
  value of a response variable.}

\note{This software program has been reviewed in accordance with
  U.S. Environmental Protection Agency policy and approved for
  publication.  Mention of trade names or commercial products does not
  constitute endorsement or recommendation for use.
}