\name{aspect.ratio.plot}
\alias{aspect.ratio.plot}

\title{
  This function is now defunct.
}


\description{This function is now defunct.  A newer version can be
  found in the "figdim" package.  See
  \url{http://www.people.vcu.edu/~jswall/figdim_index.html}.}


\usage{
aspect.ratio.plot(...)
}


\arguments{
  \item{\dots}{Since the function is now defunct, a message explaining
  this will be returned, regardless of the arguments passed.}
}


\details{This function is now defunct.  A newer version can be
  found in the "figdim" package.  See
  \url{http://www.people.vcu.edu/~jswall/figdim_index.html}.
}


\author{Jenise Swall}


\seealso{\code{\link{par}}, \code{\link{plot}}, \code{\link{image}}, \code{\link{Devices}}}



\keyword{ hplot }
\keyword{ device }
