\name{init.fig.dimen}
\alias{init.fig.dimen}

\title{Initializes a device for a figure with specified dimensions.}

\description{This function is now defunct.  A newer version can be
  found in the "figdim" package.  See
  \url{http://www.people.vcu.edu/~jswall/figdim_index.html}.}


\usage{init.fig.dimen(...)
}


\arguments{
  \item{\dots}{Since the function is now defunct, a message explaining
    this will be returned, regardless of the arguments passed. }
}


\details{This function is now defunct.  A newer version can be
  found in the "figdim" package.  See
  \url{http://www.people.vcu.edu/~jswall/figdim_index.html}.
}


\author{Jenise Swall}


\seealso{\code{\link{par}}}



\keyword{ hplot }
\keyword{ device }
